microservices/app/src/requirements.txt
request
wit
tweepy
